+++ README.md
@@ -1,4 +1,4 @@
-# simple-gRPC-app
+# Simple gRPC app
 A simple application with frontend and backend process using gRPC. 
 Backend starts a gRPC server and listens for requests. It returns the value of
 the requested env variable. 
@@ -12,24 +12,25 @@ Frontend starts a http server and renders the backend's output as html.
 
 ## Frontend
 ```
-Usage of frontend:
--addr string
-Server address string (default ":8080")
--env string
-Environment variable name
+Usage of ./frontend:
+  -addr string
+    	Server address string (default ":8080")
 ```
+NOTE: `GRPC_SERVER env variable can be used to set the address if the gRPC server is running on different node`
 
 ## Example
-1. Start frontend and backend on two different terminals
+Start frontend and backend on two different terminals
+
 ```
-stiptur@mb02287 simple-gRPC-app % ./_build/frontend -env=SHELL
+stiptur@mb02287 simple-gRPC-app % ./_build/frontend
 2023/03/22 11:16:02 Server is listening
 
 stiptur@mb02287 simple-gRPC-app % ./_build/backend
 ```
-2. Check output by using curl or browser
+Check output by using curl or browser
+
 ```
-stiptur@mb02287 simple-gRPC-app % curl http://localhost:8080/getenv
+stiptur@mb02287 simple-gRPC-app % curl http://localhost:8080/getenv\?env=SHELL
 <!DOCTYPE html>
 <html lang="en">
 
@@ -45,8 +46,3 @@ stiptur@mb02287 simple-gRPC-app % curl http://localhost:8080/getenv
 
 </html>
 '''
-
-
-
-
-
