+++ cmd/backend/main.go
@@ -1,33 +1,16 @@
 package main
 
 import (
-	"context"
-	"fmt"
 	"log"
 	"net"
-	"os"
 
+	"github.com/sachintiptur/grpc-app/pkg/backend"
 	"google.golang.org/grpc"
 
 	pb "github.com/sachintiptur/grpc-app/proto"
 )
 
-type server struct {
-	pb.UnimplementedGetterServer
-}
-
-// GetEnvVariable gets the environment variable value
-func (s *server) GetEnvVariable(ctx context.Context, req *pb.EnvRequest) (*pb.EnvResponse, error) {
-	env := req.GetEnvName()
-	if _, ok := os.LookupEnv(env); !ok {
-		return &pb.EnvResponse{}, fmt.Errorf("env %s is not set", env)
-	}
-	val := os.Getenv(env)
-	return &pb.EnvResponse{EnvValue: val}, nil
-}
-
-// backend gRPC server
-// listens to gRPC requests and return environment variable's value
+// Backend gRPC server listens to gRPC requests and return environment variable's value
 func main() {
 	lis, err := net.Listen("tcp", ":9000")
 	if err != nil {
@@ -36,7 +19,7 @@ func main() {
 
 	// register and start gRPC server
 	srv := grpc.NewServer()
-	pb.RegisterGetterServer(srv, &server{})
+	pb.RegisterGetterServer(srv, &backend.Server{})
 	err = srv.Serve(lis)
 	if err != nil {
 		log.Fatalf("failed to start server: %v", err)
