+++ proto/env.proto
@@ -1,8 +1,9 @@
 syntax = "proto3";
 
-option go_package = "github.com/sachintiptur/simple-app/proto";
+option go_package = "github.com/sachintiptur/simple-gRPC-app/proto";
 
 package env;
+
 // The request message containing environment variable name
 message EnvRequest {
   string env_name  = 1;
